Add support for multipart proxying in site replication
    This will faciliate clients sending upload part requests to the load
    balancer instead of remembering the site that received the initial
    upload. Multipart proxying will ensure upload part, complete multipart
    and abort requests are redirected to the correct site that initiated
    the upload.
