+++ cmd/bucket-replication.go
@@ -1414,7 +1414,7 @@ func replicateObjectWithMultipart(ctx context.Context, c *miniogo.Core, bucket, 
 		if err != nil {
 			return err
 		}
-		pInfo, err = c.PutObjectPart(ctx, bucket, object, uploadID, partInfo.Number, hr, partInfo.ActualSize, "", "", opts.ServerSideEncryption)
+		pInfo, _, err = c.PutObjectPart(ctx, bucket, object, uploadID, partInfo.Number, hr, partInfo.ActualSize, "", "", opts.ServerSideEncryption)
 		if err != nil {
 			return err
 		}
